+++ src/error_correction/Shor_code.jl
@@ -1,20 +1,5 @@
 #currently in the works
 
-module Codes
-abstract type Code end
-
-struct Shor_code <: Code
-end
-
-"""documents"""
-
-function rate end
-
-function logicalqubits end
-
-function physicalqubits end
-
-function codedistance end
 
 #function H(code::Shor_code) end
